+++ functions/functions-msg.php
@@ -1946,15 +1946,13 @@ function chat_msg($o_id, $u_id, $u_nick, $u_farbe, $admin, $r_id, $text, $typ) {
 		case "/freunde":
 		case "/freund":
 		case "/buddy":
-		// Fügt Freund der Freundesliste hinzu oder löscht einen Eintrag
-		
+			// Fügt Freund der Freundesliste hinzu oder löscht einen Eintrag
 			if ($u_level != "G") {
 				$privat = FALSE;
 				$text = html_parse($privat, htmlspecialchars($chatzeile[2] . " " . $chatzeile[3]))[0];
 				
 				if ($chatzeile[1]) {
-					
-					// Nach nicknamen suchen		
+					// Nach nicknamen suchen
 					$nick = nick_ergaenze($chatzeile[1], "online", 1);
 					
 					// Falls keinen Empfänger gefunden, in Benutzertabelle nachsehen
@@ -1967,7 +1965,6 @@ function chat_msg($o_id, $u_id, $u_nick, $u_farbe, $admin, $r_id, $text, $typ) {
 							$nick = mysqli_fetch_array($result, MYSQLI_ASSOC);
 						}
 						mysqli_free_result($result);
-						
 					}
 					
 					$freunddazu = 1;
@@ -1989,31 +1986,19 @@ function chat_msg($o_id, $u_id, $u_nick, $u_farbe, $admin, $r_id, $text, $typ) {
 					
 					// Falls nick gefunden, prüfen ob er bereits in Tabelle steht
 					if ($nick['u_nick'] != "" && $nick['u_id']) {
-						$query = "SELECT f_id from freunde WHERE "
-							. "(f_userid=$nick[u_id] AND f_freundid=$u_id ) "
-							. "OR "
-							. "(f_userid=$u_id AND f_freundid=$nick[u_id] )";
+						$query = "SELECT f_id from freunde WHERE (f_userid=$nick[u_id] AND f_freundid=$u_id ) OR (f_userid=$u_id AND f_freundid=$nick[u_id] )";
 						
 						$result = mysqli_query($mysqli_link, $query);
 						if ($result && mysqli_num_rows($result) > 0) {
 							
 							// Benutzer ist bereits Freund -> entfernen
-							$query = "DELETE from freunde WHERE "
-								. "(f_userid=$nick[u_id] AND f_freundid=$u_id) "
-								. "OR "
-								. "(f_userid=$u_id AND f_freundid=$nick[u_id])";
+							$query = "DELETE from freunde WHERE (f_userid=$nick[u_id] AND f_freundid=$u_id) OR (f_userid=$u_id AND f_freundid=$nick[u_id])";
 							$result = mysqli_query($mysqli_link, $query);
-							system_msg("", 0, $u_id, $system_farbe,
-								str_replace("%u_nick%", $nick['u_nick'], $t['chat_msg84']));
-							
-						} elseif ($nick['u_id'] == $u_id) {
-							
+							system_msg("", 0, $u_id, $system_farbe, str_replace("%u_nick%", $nick['u_nick'], $t['chat_msg84']));
+						} else if ($nick['u_id'] == $u_id) {
 							// Eigener Freund ist verboten
-							system_msg("", 0, $u_id, $system_farbe,
-								$t['chat_msg89']);
-							
+							system_msg("", 0, $u_id, $system_farbe, $t['chat_msg89']);
 						} else {
-							
 							if ($freunddazu == 1) {
 								// Benutzer ist noch kein Freund -> hinzufügen
 								unset($f);
@@ -2027,22 +2012,16 @@ function chat_msg($o_id, $u_id, $u_nick, $u_farbe, $admin, $r_id, $text, $typ) {
 								# system_msg("",0,$u_id,$system_farbe,str_replace("%u_nick%",$nick[u_nick],$t[chat_msg83]));
 								system_msg("", 0, $u_id, $system_farbe, $text);
 							} else if ($freunddazu == -1) {
-								system_msg("", 0, $u_id, $system_farbe,
-									str_replace("%u_nick%", $nick['u_nick'], $t['chat_msg116']));
+								system_msg("", 0, $u_id, $system_farbe, str_replace("%nickname%", $nick['u_nick'], $t['freunde_fehlermeldung_ignoriert']));
 							} else if ($freunddazu == -2) {
-								system_msg("", 0, $u_id, $system_farbe,
-									str_replace("%u_nick%", $nick['u_nick'], $t['chat_msg117']));
+								system_msg("", 0, $u_id, $system_farbe, str_replace("%nickname%", $nick['u_nick'], $t['freunde_fehlermeldung_gesperrt']));
 							} else if ($freunddazu == -3) {
-								system_msg("", 0, $u_id, $system_farbe,
-									str_replace("%u_nick%", $nick['u_nick'], $t['chat_msg118']));
+								system_msg("", 0, $u_id, $system_farbe, str_replace("%nickname%", $nick['u_nick'], $t['freunde_fehlermeldung_gast']));
 							}
-							
 						}
 					} else {
 						// Benutzer nicht gefunden
-						system_msg("", 0, $u_id, $system_farbe,
-							str_replace("%chatzeile%", $chatzeile[1],
-								$t['chat_msg8']));
+						system_msg("", 0, $u_id, $system_farbe, str_replace("%chatzeile%", $chatzeile[1], $t['chat_msg8']));
 					}
 				}
